[FrameworkBundle] Move the passes and commands that act on another package to that package
