+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 8.2
 ---
 
+ * Deprecate `Command\RouterMatchCommand`, use the one from the Routing component instead
+ * Deprecate `DependencyInjection\Compiler\TranslationLintCommandPass`, `DependencyInjection\Compiler\TranslationUpdateCommandPass`, `DependencyInjection\Compiler\AssetsContextPass` and `DependencyInjection\Compiler\AddValidatorSecurityExpressionLanguageProviderPass`, use the ones from the Translation, Asset and Validator components instead
  * Deprecate `Routing\RouteLoaderInterface`, use the `#[AsRouteLoader]` attribute from the Routing component instead
  * Deprecate `Routing\Router` and `Routing\Attribute\AsRoutingConditionService`, use the ones from the Routing component instead
  * Deprecate `Translation\Translator`, use `Symfony\Component\Translation\DependencyInjection\Translator` instead
