+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/AddValidatorSecurityExpressionLanguageProviderPass.php
@@ -11,23 +11,8 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\Validator\DependencyInjection\AddValidatorSecurityExpressionLanguageProviderPass as BaseAddValidatorSecurityExpressionLanguageProviderPass;
 
-/**
- * Registers security expression functions into the validator expression language
- * when both the Validator and Security components are available.
- */
-class AddValidatorSecurityExpressionLanguageProviderPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container): void
-    {
-        if (!$container->has('validator.expression_language') || !$container->has('security.expression_language_provider')) {
-            return;
-        }
+trigger_deprecation('symfony/framework-bundle', '8.2', 'The "%s" class is deprecated, use "%s" instead.', 'Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddValidatorSecurityExpressionLanguageProviderPass', BaseAddValidatorSecurityExpressionLanguageProviderPass::class);
 
-        $container->findDefinition('validator.expression_language')
-            ->addMethodCall('registerProvider', [new Reference('security.expression_language_provider')]);
-    }
-}
+class_alias(BaseAddValidatorSecurityExpressionLanguageProviderPass::class, 'Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddValidatorSecurityExpressionLanguageProviderPass');
