+++ src/Symfony/Component/Routing/composer.json
@@ -23,6 +23,7 @@
         "psr/container": "^1.1|^2.0",
         "psr/log": "^1|^2|^3",
         "symfony/config": "^7.4|^8.0",
+        "symfony/console": "^7.4|^8.0",
         "symfony/expression-language": "^7.4|^8.0",
         "symfony/dependency-injection": "^8.2",
         "symfony/http-foundation": "^7.4|^8.0",
