+++ src/Symfony/Component/Validator/Tests/DependencyInjection/AddValidatorSecurityExpressionLanguageProviderPassTest.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\Compiler;
+namespace Symfony\Component\Validator\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddValidatorSecurityExpressionLanguageProviderPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\Validator\DependencyInjection\AddValidatorSecurityExpressionLanguageProviderPass;
 
 class AddValidatorSecurityExpressionLanguageProviderPassTest extends TestCase
 {
